requirements.txt
flask>=2.0.0
werkzeug>=2.0.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
Pillow>=9.0.0
numpy>=1.24.0
gunicorn>=20.1.0
opencv-python-headless>=4.7.0.72
jinja2>=3.0.0
itsdangerous>=2.0.0
tqdm>=4.64.0 